+++ cmp/grouping/GroupingChooserModel.ts
@@ -274,6 +274,10 @@ export class GroupingChooserModel extends HoistModel {
         return value.map(dimName => this.getDimDisplayName(dimName)).join(' › ');
     }
 
+    getValueDisplayNames(): string[] {
+        return this.value.map(dimName => this.getDimDisplayName(dimName));
+    }
+
     getDimDisplayName(dimName: string) {
         return this.dimensions[dimName]?.displayName ?? dimName;
     }